game: mover, savegame and AI fixes
- IsBinaryMoverBlocked dereferenced a NULL activator when something
  other than a target_relay triggered a rotating door
- doors use soundLoop while moving like every other mover state; the
  old code assigned the one-shot sound2to3 and had an unreachable branch
- savegames convert gclient_t.persistantPowerup like the other entity
  pointers and ignore modelInfo (set again by ClientUserinfoChanged)
  instead of restoring raw pointers from another run
- fuzzy weight interpolation used integer division
- guard the enemy list memcpy against an empty list

+++ src/botlib/be_ai_weight.c
@@ -654,7 +654,7 @@ float FuzzyWeight_r( int *inventory, fuzzyseperator_t *fs ) {
 				w2 = FuzzyWeight_r( inventory, fs->next->child );
 			} else { w2 = fs->next->weight;}
 			//the scale factor
-			scale = ( inventory[fs->index] - fs->value ) / ( fs->next->value - fs->value );
+			scale = (float)( inventory[fs->index] - fs->value ) / ( fs->next->value - fs->value );
 			//scale between the two weights
 			return scale * w1 + ( 1 - scale ) * w2;
 		} //end if
@@ -687,7 +687,7 @@ float FuzzyWeightUndecided_r( int *inventory, fuzzyseperator_t *fs ) {
 				w2 = FuzzyWeight_r( inventory, fs->next->child );
 			} else { w2 = fs->next->minweight + random() * ( fs->next->maxweight - fs->next->minweight );}
 			//the scale factor
-			scale = ( inventory[fs->index] - fs->value ) / ( fs->next->value - fs->value );
+			scale = (float)( inventory[fs->index] - fs->value ) / ( fs->next->value - fs->value );
 			//scale between the two weights
 			return scale * w1 + ( 1 - scale ) * w2;
 		} //end if
